Removing remaining explore redux code
