+++ src/lib/redux-store/model/model-apis.ts
@@ -1,9 +1,10 @@
 import { dataModelerService } from "$lib/application-state-stores/application-store";
 import { selectMetricsDefinitionsByModelId } from "$lib/redux-store/dimension-definition/dimension-definition-selectors";
 import { store } from "$lib/redux-store/store-root";
-import { setExplorerIsStale } from "$lib/redux-store/explore/explore-slice";
 import { selectApplicationActiveEntity } from "$lib/redux-store/application/application-selectors";
 import { validateSelectedSources } from "$lib/redux-store/metrics-definition/metrics-definition-apis";
+import { invalidateMetricView } from "$lib/svelte-query/queries/metric-view";
+import { queryClient } from "$lib/svelte-query/globalQueryClient";
 
 export const updateModelQueryApi = async (
   modelId: string,
@@ -30,7 +31,7 @@ const syncMetricsDefinitions = (modelId: string) => {
 
   const metricsDefinitions = selectMetricsDefinitionsByModelId(state, modelId);
   metricsDefinitions.forEach((metricsDefinition) => {
-    store.dispatch(setExplorerIsStale(metricsDefinition.id, true));
+    invalidateMetricView(queryClient, metricsDefinition.id);
     if (activeEntity.id === metricsDefinition.id) {
       store.dispatch(validateSelectedSources(metricsDefinition.id));
     }
