+++ src/routes/__layout.svelte
@@ -15,10 +15,14 @@
   import { initMetrics } from "$lib/metrics/initMetrics";
   import { syncApplicationState } from "$lib/redux-store/application/application-apis";
   import type { ApplicationMetadata } from "$lib/types";
-  import { QueryClient, QueryClientProvider } from "@sveltestack/svelte-query";
+  import { QueryClientProvider } from "@sveltestack/svelte-query";
   import { onMount, setContext } from "svelte";
   import "../app.css";
   import "../fonts.css";
+  import {
+    createQueryClient,
+    queryClient,
+  } from "$lib/svelte-query/globalQueryClient.js";
 
   let store;
   let queryHighlight = createQueryHighlightStore();
@@ -42,14 +46,7 @@
     syncApplicationState(store);
   }
 
-  const queryClient = new QueryClient({
-    defaultOptions: {
-      queries: {
-        refetchOnWindowFocus: false,
-        placeholderData: {},
-      },
-    },
-  });
+  createQueryClient();
 
   onMount(() => {
     initMetrics();
