+++ src/routes/_surfaces/workspace/explore/Explore.svelte
@@ -1,17 +1,10 @@
 <script lang="ts">
-  import { bootstrapMetricsExplorer } from "$lib/redux-store/explore/bootstrapMetricsExplorer";
-  import { store } from "$lib/redux-store/store-root";
-  import { onMount } from "svelte";
   import ExploreContainer from "./ExploreContainer.svelte";
   import ExploreHeader from "./ExploreHeader.svelte";
   import LeaderboardDisplay from "./leaderboards/LeaderboardDisplay.svelte";
   import MetricsTimeSeriesCharts from "./time-series-charts/MetricsTimeSeriesCharts.svelte";
 
   export let metricsDefId: string;
-
-  onMount(() => {
-    store.dispatch(bootstrapMetricsExplorer(metricsDefId));
-  });
 
   let whichReferenceValue: string;
 </script>
