+++ src/routes/_surfaces/workspace/explore/filters/Filters.svelte
@@ -1,9 +1,11 @@
 <script lang="ts">
   import type { MetricViewRequestFilter } from "$common/rill-developer-service/MetricViewActions";
 
-  import { metricsExplorerStore } from "$lib/application-state-stores/explorer-stores";
+  import {
+    MetricsExplorerEntity,
+    metricsExplorerStore,
+  } from "$lib/application-state-stores/explorer-stores";
   import Close from "$lib/components/icons/Close.svelte";
-  import type { MetricsExplorerEntity } from "$lib/redux-store/explore/explore-slice";
   import { invalidateMetricViewData } from "$lib/svelte-query/queries/metric-view";
   import { useQueryClient } from "@sveltestack/svelte-query";
   import { fly } from "svelte/transition";
