+++ src/routes/index.svelte
@@ -1,9 +1,6 @@
 <script lang="ts">
   import { EntityStatus } from "$common/data-modeler-state-service/entity-state-service/EntityStateService";
-  import {
-    ApplicationStore,
-    config,
-  } from "$lib/application-state-stores/application-store";
+  import { ApplicationStore } from "$lib/application-state-stores/application-store";
   import {
     assetsVisible,
     assetVisibilityTween,
@@ -31,8 +28,6 @@
   import PreparingImport from "$lib/components/overlay/PreparingImport.svelte";
   import QuickStartDashboard from "$lib/components/overlay/QuickStartDashboard.svelte";
   import SurfaceControlButton from "$lib/components/surface/SurfaceControlButton.svelte";
-  import { HttpStreamClient } from "$lib/http-client/HttpStreamClient";
-  import { store } from "$lib/redux-store/store-root";
   import { getContext } from "svelte";
   import AssetsSidebar from "./_surfaces/assets/index.svelte";
   import InspectorSidebar from "./_surfaces/inspector/index.svelte";
@@ -69,8 +64,6 @@
   $: persistentExportedModel = $persistentModelStore?.entities?.find(
     (model) => model.id === derivedExportedModel?.id
   );
-
-  HttpStreamClient.create(`${config.server.serverUrl}/api`, store.dispatch);
 
   /** Workaround for hiding inspector for now. Post July 19 2022 we will remove this
    * in favor of ironing out more modular routing and suface management.
